requirements.txt
streamlit
langchain
faiss-cpu
ollama



